Workaround Arch's block break not being sufficient
The BlockEvent.BREAK event from Arch seems to not fire for PlayerBlockBreakEvents.BEFORE which Fabric mods uses for checking if their actions are allowed or not by claims mods.

This PR does a simple solution of registering to the PlayerBlockBreakEvents.BEFORE event on the Fabric side manually. Should be good for compat and work with my Bumblezone mod (I tested in dev with this change and it looks all good)